+++ CHANGELOG.md
@@ -5,10 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for `argoproj.io/v1alpha1` `Rollout` resource.
 ### Changed
 - Add clarity between `autoscaling/v1` and `autoscaling/v2beta2` APIs, no breaking changes.
 ### Security
-- Updated k8s client libraries to `v1.22.5` to resolve
+- Updated k8s client libraries to `v0.21.8` to resolve
 [GHSA-f5f7-6478-qm6p](https://github.com/advisories/GHSA-f5f7-6478-qm6p).
 
 ## [v2.2.0] - 2021-08-14